pyproject.toml
[project]
name = "bank"
version = "0.1.0"
dependencies = [
  "fastapi>=0.112.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "psycopg2>=2.9.9",
  "faker>=28.4.1",
  "unitest>=1.4.34"
]
requires-python = ">=3.11"

description = "bank"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "mypy",
    "pytest",
    "pytest-sqlalchemy"

]